'Consolidated the 8ball command from the memes cog, the coin flip cog, and the dice roll cog all into one cog folder known as random_cog'
