+++ bot/cogs/random_cog/assets/Heads.jpg
