+++ bot/cogs/random_cog/assets/Tails.jpg
